backend/requirements.txt
flask
flask-cors
yt-dlp
google-cloud-speech
google-generativeai
